moved acomponent to  pages directory
